Default to clang instead of GCC for compiling simple example
Ultimately I am sick of dealing with the various cross-compilers for GCC
and the different behaviour between cross-compilers that are supposed to
be doing the same thing (e.g. aarch64-none-elf and aarch64-linux-gnu).
This problem will only be worse when we add support for more architectures
to the VMM as every architecture is a separate set of binaries needed to
be installed.

With Clang and LLD, there are only *two* binaries needed for *all*
targets we will support. This is significantly easier to mantain and
less likely to result in weird toolchain bugs unlike with the various
GCC cross-compilers for the same target. If we were to use Zig's packaged
toolchain, we could even get away with having a single binary for compiling
all of our code.

Of course, we still want to support using GCC to compile the library
so after this patch we will still need to test that GCC works.