Disable hardening flags in Nix shell
One of the flags being passed to the C compiler automatically by Nix
`-fstack-protector-strong --param ssp-buffer-size=4`, assumes that
we have a some functionality available in libc which does not exist
when targeting a freestanding target.

For good measure and having the Nix shell environment be closer to a
native environment, we disable all of these hardening flags.