Remove Raspberry Pi 4B+ images until they are in a working state
The support for this platform was not fully completed and what we'll
do instead is have the work-in-progress on a separate branch if we
intend to finish the port.

The reason for the removal is that we do not want to include anything
in the repository that does not work fully.