Remove support for Odroid-C2
The Odroid-C2 is actually no longer available for purchase
and is pretty much the same as the Odroid-C4 (which is available)
such that having the Odroid-C2 support in the VMM examples is of
limited use.

However, if you are reading this commit message and want Odroid-C2
support, please contact the maintainers of the project. It should
be trivial to add support back to the simple example.