Fix UB in HSR exception class decoding
The Zig packaged C compiler caught this UB as it, by default,
turns on LLVM's UBSAN checking. This caused a `brk` instruction
to be inserted such that when running the VMM built via Zig
meant that it would cause an exception.

Neither GCC or Clang had this behaviour.

All of this HSR code and vCPU fault handling should be revisited
and audited in the near future.