Fixes for latest master Zig version
There are some minor breaking API changes to the Zig build system. It
is good to get ahead of these so we don't have to deal with all of them
at the next release of Zig.

Hopefully soon we can pin the examples that use Zig to an official
release rather than the master version.