examples/simple: disable UBSAN when building with Zig
Zig enables UBSAN on C code by default and it seems to be creating
weird behaviour. If UB is detected at run-time, the program should
panic by executing a `brk` instruction in this case (ARM64 and no OS).

It seems that the Microkit event loop is actually doing different
things depending on whether or not UBSAN is on. Specifically, the
`seL4_NBSendRecv` system call is being made when UBSAN is on, even
though it should never be called in this example code (and is not
called when we turn UBSAN off).

Not sure what's going on, it's still entirely possible that we have
UB in our code. More investigation required.