Restructuring the Zig build system
This gives libvmm a top-level build.zig which means that projects that
depend on libvmm can simply add libvmm as a dependency in their
build.zig.zon file and do not have to figure out how to build libvmm.a.

This same approach is used by some of the examples.