Remove 'framebuffer' and 'ethernet' examples
These examples are unmaintained and not tested in CI. What the
examples show is worthwhile however and is planend to put in a
'passthrough' example.

I don't know when we will get to creating that passthrough example,
so in the interest of not having bitrotted examples, we remove them.

They are of course in the commit history if we ever want to revive
them.

Signed-off-by: Ivan Velickovic <i.velickovic@unsw.edu.au>