examples/simple: utilise overlays
For QEMU virt AArch64 and Odroid-C4, what we did in the past is take
the original DTS intended for booting Linux natively and adapt it to
work in a virtualised environment.

This patch gets those two platforms in line with the current convention
for other platforms and examples which is to have `linux.dts` be the
unmodified DTS from Linux source and then use the overlay to disable
certain devices, change the amount and address of RAM, etc.

Signed-off-by: Ivan Velickovic <i.velickovic@unsw.edu.au>