Remove unused struct fields in GIC register maps
These are emulated devices, and we aren't emulating the reserved
regions. This only makes sense in a context where you have a
one-to-one mapping between struct field offset and device register
offset, which we do not.

On GICv2 this saves 668 bytes, on GICv3 it saves 119368 bytes.

I'm sure there's much more to do, but this is very low-hanging
fruit.

Signed-off-by: Ivan-Velickovic <i.velickovic@unsw.edu.au>